+++ cli/src/main.rs
@@ -48,7 +48,7 @@ impl Context {
 /// Reads config files for writing data, or promps the user if they don't yet exist
 fn set_agent_config() -> CLIResult<Config> {
     let agent_config_path = atomic_lib::config::default_config_file_path()?;
-    match atomic_lib::config::read_config(&agent_config_path) {
+    match atomic_lib::config::read_config(Some(&agent_config_path)) {
         Ok(found) => Ok(found),
         Err(_e) => {
             println!(