+++ CHANGELOG.md
@@ -5,13 +5,15 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 **Changes to JS assets (including the front-end and JS libraries) are not shown here**, but in [`/browser/CHANGELOG`](/browser/CHANGELOG.md).
 See [STATUS.md](server/STATUS.md) to learn more about which features will remain stable.
 
-## [v0.36.2] - 2023-12-20
+## [v0.36.2] - 2024-01-05
 
 - Use `musl` + `alpine` builds for docker images, way smaller images #620
 - Support multi-platform docker builds #731
 - Remove deprecated ENV vars #732
 - Fix no Agent as drive
 - Add `clear` option to error component (resets all front-end state)
+- Add `Agent::from_secret` #785
+- Make `set_propval` and `set_propval_shortname` chainable #785
 
 ## [v0.36.1] - 2023-12-06